Set page to 1 in search when changing searchparams
This is to fix a bug. Previously for example you have 100 searchresults
and you are on the last page, page 10. Then you add a facet to only show
results for year 1790 for example. Then you have one result but the
paging is still on page 10 and you see no results